chore(quality): pay down Revisor quality debt and enforce CI gates
Resolve all findings from the Revisor QA audit (PR#1, run #188):
- ruff: 0 errors repo-wide (auto-fixes plus magic-value constants,
  pathlib conversions, B904 raise-from, duplicate endpoint rename,
  cache-clear via .clear(), Odoo framework per-file ignores)
- mypy: 0 errors in 56 source files (pydantic v2 Field env cleanup,
  protocol variance typevars, Optional narrowing, Awaitable retry
  retype, ErrorCode enums, type stub packages)
- bandit: 0 findings (MD5 usedforsecurity=False, justified nosec for
  local pickle and container bind-all, B110 try/except/pass converted
  to debug logging); pin bandit>=1.9 for Python 3.14 compatibility
- Latent bugs fixed: nonexistent _add_derived_features call, circuit
  breaker OPEN-to-HALF_OPEN None return, LoggingManager._local crash,
  ModelValidator predict_15_day_period typo
- CI: replace light baseline with quality-gates job (ruff format and
  lint, mypy, bandit, pytest) on push and PR
- Dev env: reinstall package as editable to fix stale 1.2.1 install
  (resolves test import failures); add legacy src.core shim

+++ src/prediction/weather_ranking_system.py
@@ -311,7 +311,7 @@ class WeatherRankingSystem:
                     hourly_scores.append(score)
 
                 # Daily average score and ranking
-                daily_score = np.mean(hourly_scores) if hourly_scores else 50.0
+                daily_score = float(np.mean(hourly_scores)) if hourly_scores else 50.0
                 daily_ranking = self._score_to_ranking(daily_score)
 
                 # Collect daily statistics
